Move 'vendor/Soup' to 'vendor/Soup/soup' to match Soup's structure
Update Soup update script

Move 'vendor/Soup/soup/LICENCE' to 'vendor/Soup/LICENCE'

Update includes

Fix Sun build

Fix Visual Studio build

Fix PHP scripts build

Fix Make build
