+++ Pluto.vcxproj
@@ -475,22 +475,21 @@
     <ClCompile Include="src\lvector3lib.cpp" />
     <ClCompile Include="src\lvm.cpp" />
     <ClCompile Include="src\lzio.cpp" />
-    <ClCompile Include="src\vendor\Soup\base32.cpp" />
-    <ClCompile Include="src\vendor\Soup\base64.cpp" />
-    <ClCompile Include="src\vendor\Soup\json.cpp" />
-    <ClCompile Include="src\vendor\Soup\JsonArray.cpp" />
-    <ClCompile Include="src\vendor\Soup\JsonBool.cpp" />
-    <ClCompile Include="src\vendor\Soup\JsonFloat.cpp" />
-    <ClCompile Include="src\vendor\Soup\JsonInt.cpp" />
-    <ClCompile Include="src\vendor\Soup\JsonNode.cpp" />
-    <ClCompile Include="src\vendor\Soup\JsonNull.cpp" />
-    <ClCompile Include="src\vendor\Soup\JsonObject.cpp" />
-    <ClCompile Include="src\vendor\Soup\JsonString.cpp" />
-    <ClCompile Include="src\vendor\Soup\Optional.hpp" />
-    <ClCompile Include="src\vendor\Soup\spaceship.cpp" />
-    <ClCompile Include="src\vendor\Soup\unicode.cpp" />
-    <ClCompile Include="src\vendor\Soup\urlenc.cpp" />
-    <ClCompile Include="src\vendor\Soup\version_compare.cpp" />
+    <ClCompile Include="src\vendor\Soup\soup\base32.cpp" />
+    <ClCompile Include="src\vendor\Soup\soup\base64.cpp" />
+    <ClCompile Include="src\vendor\Soup\soup\json.cpp" />
+    <ClCompile Include="src\vendor\Soup\soup\JsonArray.cpp" />
+    <ClCompile Include="src\vendor\Soup\soup\JsonBool.cpp" />
+    <ClCompile Include="src\vendor\Soup\soup\JsonFloat.cpp" />
+    <ClCompile Include="src\vendor\Soup\soup\JsonInt.cpp" />
+    <ClCompile Include="src\vendor\Soup\soup\JsonNode.cpp" />
+    <ClCompile Include="src\vendor\Soup\soup\JsonNull.cpp" />
+    <ClCompile Include="src\vendor\Soup\soup\JsonObject.cpp" />
+    <ClCompile Include="src\vendor\Soup\soup\JsonString.cpp" />
+    <ClCompile Include="src\vendor\Soup\soup\spaceship.cpp" />
+    <ClCompile Include="src\vendor\Soup\soup\unicode.cpp" />
+    <ClCompile Include="src\vendor\Soup\soup\urlenc.cpp" />
+    <ClCompile Include="src\vendor\Soup\soup\version_compare.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\lerrormessage.hpp" />
@@ -526,35 +525,36 @@
     <ClInclude Include="src\lundump.h" />
     <ClInclude Include="src\lvm.h" />
     <ClInclude Include="src\lzio.h" />
-    <ClInclude Include="src\vendor\Soup\base.hpp" />
-    <ClInclude Include="src\vendor\Soup\base32.hpp" />
-    <ClInclude Include="src\vendor\Soup\base64.hpp" />
-    <ClInclude Include="src\vendor\Soup\bitutil.hpp" />
-    <ClInclude Include="src\vendor\Soup\Endian.hpp" />
-    <ClInclude Include="src\vendor\Soup\Exception.hpp" />
-    <ClInclude Include="src\vendor\Soup\fwd.hpp" />
-    <ClInclude Include="src\vendor\Soup\IntStruct.hpp" />
-    <ClInclude Include="src\vendor\Soup\ioBase.hpp" />
-    <ClInclude Include="src\vendor\Soup\json.hpp" />
-    <ClInclude Include="src\vendor\Soup\JsonArray.hpp" />
-    <ClInclude Include="src\vendor\Soup\JsonBool.hpp" />
-    <ClInclude Include="src\vendor\Soup\JsonFloat.hpp" />
-    <ClInclude Include="src\vendor\Soup\JsonInt.hpp" />
-    <ClInclude Include="src\vendor\Soup\JsonNode.hpp" />
-    <ClInclude Include="src\vendor\Soup\JsonNodeType.hpp" />
-    <ClInclude Include="src\vendor\Soup\JsonNull.hpp" />
-    <ClInclude Include="src\vendor\Soup\JsonObject.hpp" />
-    <ClInclude Include="src\vendor\Soup\JsonString.hpp" />
-    <ClInclude Include="src\vendor\Soup\Reader.hpp" />
-    <ClInclude Include="src\vendor\Soup\spaceship.hpp" />
-    <ClInclude Include="src\vendor\Soup\string.hpp" />
-    <ClInclude Include="src\vendor\Soup\TreeNode.hpp" />
-    <ClInclude Include="src\vendor\Soup\type_traits.hpp" />
-    <ClInclude Include="src\vendor\Soup\unicode.hpp" />
-    <ClInclude Include="src\vendor\Soup\UniquePtr.hpp" />
-    <ClInclude Include="src\vendor\Soup\urlenc.hpp" />
-    <ClInclude Include="src\vendor\Soup\version_compare.hpp" />
-    <ClInclude Include="src\vendor\Soup\Writer.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\base.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\base32.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\base64.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\bitutil.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\Endian.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\Exception.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\fwd.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\IntStruct.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\ioBase.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\json.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\JsonArray.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\JsonBool.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\JsonFloat.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\JsonInt.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\JsonNode.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\JsonNodeType.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\JsonNull.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\JsonObject.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\JsonString.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\Optional.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\Reader.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\spaceship.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\string.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\TreeNode.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\type_traits.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\unicode.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\UniquePtr.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\urlenc.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\version_compare.hpp" />
+    <ClInclude Include="src\vendor\Soup\soup\Writer.hpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
