+++ src/lbase32.cpp
@@ -3,7 +3,7 @@
 #include "lauxlib.h"
 #include "lualib.h"
 
-#include "vendor/Soup/base32.hpp"
+#include "vendor/Soup/soup/base32.hpp"
 
 static int encode(lua_State* L) {
 	lua_pushstring(L, soup::base32::encode(luaL_checkstring(L, 1), (bool)lua_toboolean(L, 2)).c_str());
