+++ src/lbaselib.cpp
@@ -23,7 +23,7 @@
 #include "lobject.h"
 #include "lstate.h"
 
-#include "vendor/Soup/version_compare.hpp"
+#include "vendor/Soup/soup/version_compare.hpp"
 
 
 static int luaB_print (lua_State *L) {
