+++ src/ljson.hpp
@@ -2,15 +2,15 @@
 
 // https://github.com/calamity-inc/Soup-Lua-Bindings/blob/main/soup_lua_bindings.hpp
 
-#include "vendor/Soup/json.hpp"
-#include "vendor/Soup/JsonInt.hpp"
-#include "vendor/Soup/JsonBool.hpp"
-#include "vendor/Soup/JsonNode.hpp"
-#include "vendor/Soup/JsonFloat.hpp"
-#include "vendor/Soup/JsonArray.hpp"
-#include "vendor/Soup/JsonObject.hpp"
-#include "vendor/Soup/JsonString.hpp"
-#include "vendor/Soup/UniquePtr.hpp"
+#include "vendor/Soup/soup/json.hpp"
+#include "vendor/Soup/soup/JsonInt.hpp"
+#include "vendor/Soup/soup/JsonBool.hpp"
+#include "vendor/Soup/soup/JsonNode.hpp"
+#include "vendor/Soup/soup/JsonFloat.hpp"
+#include "vendor/Soup/soup/JsonArray.hpp"
+#include "vendor/Soup/soup/JsonObject.hpp"
+#include "vendor/Soup/soup/JsonString.hpp"
+#include "vendor/Soup/soup/UniquePtr.hpp"
 
 static bool isIndexBasedTable(lua_State* L, int i)
 {
