+++ src/lparser.cpp
@@ -42,7 +42,7 @@
 
 #include "lerrormessage.hpp"
 
-#include "vendor/Soup/version_compare.hpp"
+#include "vendor/Soup/soup/version_compare.hpp"
